move modules into package dir
