+++ schematic_db/__init__.py
@@ -1,0 +1,2 @@
+"""Main package"""
+from .db_object_config import *
