+++ schematic_db/rdb_queryer/rdb_queryer.py
@@ -1,7 +1,7 @@
 """RDB Queryer"""
 import pandas as pd
-from rdb import RelationalDatabase
-from query_store import QueryStore
+from schematic_db.rdb import RelationalDatabase
+from schematic_db.query_store import QueryStore
 
 
 class DuplicateColumnError(Exception):
