+++ tests/test_mysql.py
@@ -17,9 +17,9 @@ will be used.
 from datetime import datetime
 import pytest
 import pandas as pd
-from db_object_config.db_object_config import DBObjectConfig
+from schematic_db.db_object_config.db_object_config import DBObjectConfig
 
-from rdb.mysql import MySQLDatabase
+from schematic_db.rdb.mysql import MySQLDatabase
 
 
 @pytest.mark.fast
